pyproject.toml
[build-system]
requires = ["setuptools>=61.2"]
build-backend = "setuptools.build_meta"


[project]
name = "cleedpy"
dynamic = ["version"]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: C",
    "Programming Language :: Python",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

requires-python = ">=3.10"
dependencies = [
    "click",
    "numpy",
    "pyyaml",
    "scipy",
]

[project.optional-dependencies]
dev = [
  "bumpver==2023.1129",
  "pre-commit==3.6.0",
  "pytest==7.4.4",
  "pytest-cov==4.1.0",
]

[project.scripts]
cleedpy-rfactor = "cleedpy.cli.rfactor:cli"
cleedpy-search = "cleedpy.cli.search:cli"

[tool.setuptools]
include-package-data = true
zip-safe = false
license-files = ["LICENSE"]

[tool.setuptools.packages]
find = {namespaces = false}

[tool.setuptools.dynamic]
version = {attr = "cleedpy.version.__version__"}
